Add, define and manage errors in our form - on the client side and on the server side
Handling Errors in Express Apps
Add a validation middleware, on the client side, using joi, add an error handler, define an ExpressError class, and define an error template for the form. Create the utils folder, with the ExpressError, the catchAsync function, and the joi schemas.
Tags: joi, catchAsync, HTTP Status Codes