Add the review's model
Create a new model for the reviews and add it to the views area.
Apply some styling and add a form and button for now.

This will be displayed in each area, and when you delete an area,
you also delete the reviews associated with that area,
in the database and on the area's front page.
Need to change the input range of the rating reviews,
to a more discreet one, with stars, for example.