Add author and authorization
Add an author to the groundwork model.
Add authorization to edit or delete an area based on the specific author of the groundwork area.
Add the validation middleware for the author and move the other validation middleware modules to their own files in the utils folder.
Import and use the middleware modules in their respective routes.