Add Starability Rating Stars
Add the css file for the stars, called stars.css,
with the source code from Starability.
Add a link to the css source in show.ejs file and
insert the html structure into the page.
Update the rating entry in the reviews view,
adding stars with a growing and rotating animation.

Tags: Starability