Add image upload
Add the .env file and save the credentials for Cloudinary and the
database keys.
Add the Multer middleware to parse the upload images in the Areas field.
Install Cloudinary and its configuration file to manage and manipulate
images, and upload them to the Cloudinary storage service.
Add an image schema, with URL and filename as keys, both of type string.
Add these two keys in a single array to the GroundworkSchema.
Store uploaded image links in Mongodb after requesting their
filenames and URLs in array form. Then map over the array that
has been added to req.files thanks to Multer. Then save the new area.
Display images with a slideshow from bootstrap.
Fix seeds to display uploaded images.
Add the enctype attribute with the value of "multipart/form-data" to
upload files in edit view. And add the reference key name="image"
to use with the upload method of multer in to the edit route.
Request these files in the form of an array in the UpdateArea method
on the areaController. Spread this array of objects into a variable
before pushing it to the images of the area.
To delete images from Mongodb, use the updateOne method to pull the
elements out of the array on the image array, where the filename of
each image is in req.body.deleteImages array, but only do this
if there's an image to delete.
To delete the image also from Cloudinary we import Cloudinary
into the controller. If there's an image to delete, loop over the
req.body.deleteImages array, and use the method from Cloudinary
Cloudinary.upload.destroy(filename) and after that update the area.
Add the Cloudinary function to request a modified version of our images,
in the form of thumbnails in the editing area,
to save time when requesting these resources.
Add the Mongoose virtual property on each image in the
ImageSchema model to set the version modified by the Cloudinary
function and display it in the edit view.