Add maps from Mapbox
Import Mapbox and add Mapbox GL JS links and scripts to the boilerplate.
Create different files in the public/js folder for showPageMap and
for clusterMap, import the script, the mapToken variable, and the
converted JSON of the area model to the index.ejs and show.ejs views.
Pre-add the access token for the map to the .env variables.
Add to the groundwork model the geometric coordinates and type properties, to set the correct properties for our map.
And access the latitude and longitude of the cities from our
seed cities.js file.
Add a new map with its properties to the showPageMap file, add a
marker and set a popup with area title and location,
set the coordinates and then add the map.
In the areas view, add the div. with the map as id and the width and
height to display the map.
In clusterMap.js add the cluster map but change the properties.
Change the source of the data to match our groundwork object.
Add the custom virtual property on the groundwork model. Add the properties name required by the map features variable, add to this the popUpMarkup property to include in our clusterMap.
Because mongoose doesn't include virtual when we convert a document to json, to include the JavaScript tokens and objects in the EJS templates, set the option toJson to enable virtual and use it in the map of the index view.