Refactor styles, and add security updates
Clean up styles and fix security issues.
Refactor the styling of the login and registration.
This includes adding the view for the home page and including
new style sheets for our views. Add map controls on both maps.
Add sanitization for the Mongodb database to protect against malicious scripts that contain common Mongodb operators with mongo-sanitize.
Add sanitization for html inputs with sanitize-html, add a custom
function to the Joi schema on each schema string.
Add the Helmet package to add additional protection and protect
the routes of the apis used in the project.
Add new configuration for our cookie-session and save it to the
Mongodb database.